Break references when adding elements
I've decided to change the behavior of array_part() with respect to
references. Now, when adding elements to the result, if the element was
originally a reference, it will be separated and made a non-reference.

The previous behavior was inconsistent as:
* It was impossible to enforce in cases like:
  $f = 1; $arg = [&$f];
  $arg2 = array_part($arg, [0]);
  without returning by reference.
* References would not be preserved in non-last level elements (arrays),
  even if we were keeping all the elements.

References in levels deeper than array_part() traverses are not affected
and are kept.